+++ server/src/exchange/exchangeRouter.ts
@@ -1,6 +1,6 @@
 import express from 'express'
 import { NextFunction, Request, Response } from "express";
-import { currencyParser } from '../routeParsers'
+import { currencyParser } from '../routing/routeParsers'
 
 export const exchangeRouter = express()
